+++ CHANGELOG.md
@@ -12,11 +12,26 @@ int func_test(void) {
   return 1==1 && 2==2 && 3==3;
 }
 ```
-* パスワードを変えられないと。
-
 * 問題修正 ... test コラムを有効に。
 * /admin/users ... ユーザを一覧表示し、パスワードを初期化、エントリーを削除する。
 * /admin/comments ... 何をする予定だったっけ？
+* コピペ予防 css
+* restrict 10 answers a day
+
+## 0.6.3 - 2021-10-12
+### Added
+* db/answers-by-date ... 何日に回答が何件届いたか？
+* db/answers-bu-date-login ... ユーザ login は何日に回答を何件寄せたか？
+
+## 0.6.2 - 2021-10-12
+### Added
+* /ch-pass パスワードを変えられる。
+
+## 0.6.1 - 2021-10-12
+### Changed
+* improve validate-answer and create-answer!
+### Added
+* version 表示 in /about. div 要素だけまとめて左寄せしたい。
 
 ## 0.6.0 - 2021-10-11
 ### Added